USB driver for AutoLink OBD-II scan tools
This is the USB driver for AutoLink AL319, AL419, AL439, AL519, and AL539 OBD-II scan tools.  This driver also work for the Harbor Frieght scan tools, check to make sure yours is compatible with one of these.  If you actually have a code scanner, it should have had a driver provided.  Please use the one provided as this one may void your warranty and I'm not fixing or replacing your scan tool.